Allow installing to $DESTDIR
Staged Installs require a custom installation location.

This is helpful when packaging packages for release, as the tools
need to have installable files "installed" to a separate clean
subdirectory, from which they can then, well, be packaged.

`mkdir` calls are required to anticipate those subdirectories
in the temporary root not yet existing. No folder structure
necessarily exists there at the start of all this.

https://www.gnu.org/prep/standards/html_node/DESTDIR.html