Rename buildable executable to `rvrrpd`
Although a default name of `main` works, it can be very beneficial
to build executables with their final names wherever possible. An
example for where this has an impact is if during the build stages
automatic manpage generation were to be done on the executable - some
tools utilize the executable name for context.

This does however mean anyone working with the local executables
without installing them would have to adjust their tooling in use
to accommodate this rename (s/main/rvrrpd/).