Update LICENSE to allow OpenSSL
We split the full GPL v3 LICENSE text into a separate file,
called LICENSE-GPLv3. The base LICENSE file now contains the
notice itself, that is intended to be shipped as part of code.

Additionally, a section 7 clause for OpenSSL has been added,
in order to comply with requirement to allow linking to it.
This is to ensure compliance when packaging. The OpenSSL licensing
text is additionally included in the source code repository
under the LICENSE-OpenSSL name.