Set up manual page generation
For purposes of packaging the tooling, manual pages are a must.

The Makefile build steps have been adjusted to generate manpages using
the `help2man` tool which executes the locally built executables with
the `--help` and `--version` flags. The Makefile install steps have
further been adjusted to place these generated manpages in the
/usr/share/man directory by default.

This ensures documentation in manpages does not have to be maintained
by hand separated from the codebase, and is instead automatically
generated from the latest 'source of truth'.

When building locally without installation steps, you can use
the command `man -l rvrrpd.8.gz` to just 'preview' the manpage.